requirements.txt
flask==3.1.0
flask-cors==5.0.1
langchain>=0.3.20
langchain-google-genai>=2.0.11
langchain-pinecone>=0.2.3
pinecone>=5.4.0
pymupdf>=1.25.3
beautifulsoup4>=4.12.2
groq>=0.18.0
python-dotenv>=1.0.0
requests>=2.32.3
faiss-cpu>=1.10.0